+++ script/summit/install.sh
@@ -32,6 +32,7 @@ export CPATH=$ROOT/include:\$CPATH
 export LIBRARY_PATH=$ROOT/lib:\$LIBRARY_PATH
 export CC=\${OLCF_GCC_ROOT}/bin/gcc
 export CXX=\${OLCF_GCC_ROOT}/bin/g++
+export LDSHARED="\${OLCF_GCC_ROOT}/bin/gcc -shared"
 export VIRTUAL_ENV=$ROOT
 EOL
 
@@ -43,67 +44,67 @@ cd /tmp/$USER-glotzerlab-software
 python3 -m pip install --progress-bar off --no-deps --no-binary :all: cython mpi4py six numpy tables numexpr deprecation
 
 # TBB
-curl -sSLO https://github.com/01org/tbb/archive/2019_U3.tar.gz \
-    && echo "b2244147bc8159cdd8f06a38afeb42f3237d3fc822555499d7ccfbd4b86f8ece  2019_U3.tar.gz" | sha256sum -c - \
-    && tar -xzf 2019_U3.tar.gz -C . \
-    && cd tbb-2019_U3 \
+curl -sSLO https://github.com/01org/tbb/archive/2019_U8.tar.gz \
+    && echo "7b1fd8caea14be72ae4175896510bf99c809cd7031306a1917565e6de7382fba  2019_U8.tar.gz" | sha256sum -c - \
+    && tar -xzf 2019_U8.tar.gz -C . \
+    && cd tbb-2019_U8 \
     && make \
     && install -d $ROOT/lib \
     && install -m755 build/linux_*release/*.so* ${ROOT}/lib \
     && install -d $ROOT/include \
     && cp -a include/tbb $ROOT/include \
     && cd .. \
-    && rm -rf tbb-2019_U3 \
-    && rm 2019_U3.tar.gz \
+    && rm -rf tbb-2019_U8 \
+    && rm 2019_U8.tar.gz \
     || exit 1
 
 # embree is not available for power9
 
 # scipy
-curl -sSLO https://github.com/scipy/scipy/releases/download/v1.2.0/scipy-1.2.0.tar.gz \
-    && echo "51a2424c8ed80e60bdb9a896806e7adaf24a58253b326fbad10f80a6d06f2214  scipy-1.2.0.tar.gz" | sha256sum -c - \
-    && tar -xzf scipy-1.2.0.tar.gz -C . \
-    && cd scipy-1.2.0 \
+curl -sSLO https://github.com/scipy/scipy/releases/download/v1.3.1/scipy-1.3.1.tar.gz \
+    && echo "2643cfb46d97b7797d1dbdb6f3c23fe3402904e3c90e6facfe6a9b98d808c1b5  scipy-1.3.1.tar.gz" | sha256sum -c - \
+    && tar -xzf scipy-1.3.1.tar.gz -C . \
+    && cd scipy-1.3.1 \
     && LAPACK=${OLCF_NETLIB_LAPACK_ROOT}/lib64/liblapack.so BLAS=${OLCF_NETLIB_LAPACK_ROOT}/lib64/libblas.so python3 setup.py install \
-    && rm -rf scipy-1.2.0 \
+    && rm -rf scipy-1.3.1 \
     || exit 1
 
 
 
- curl -sSLO https://glotzerlab.engin.umich.edu/Downloads/freud/freud-v1.2.1.tar.gz \
-    && echo "4234ba0304742a8a2dd70df0663c23485c9cdd415789d34a434a76c116212d2e  freud-v1.2.1.tar.gz" | sha256sum -c - \
-    && tar -xzf freud-v1.2.1.tar.gz -C . \
-    && rm -f freud-v1.2.1/*.toml \
+ curl -sSLO https://glotzerlab.engin.umich.edu/Downloads/freud/freud-v2.0.0.tar.gz \
+    && echo "29c6eef2c173c04a14d773ca3b0aaba5b537b2f311d5ccd0b12db218c9cf0239  freud-v2.0.0.tar.gz" | sha256sum -c - \
+    && tar -xzf freud-v2.0.0.tar.gz -C . \
+    && rm -f freud-v2.0.0/*.toml \
     && export CFLAGS="-mcpu=power9 -mtune=power9" CXXFLAGS="-mcpu=power9 -mtune=power9" \
-    && python3 -m pip install --no-deps --ignore-installed ./freud-v1.2.1 \
-    && rm -rf freud-v1.2.1 \
-    && rm freud-v1.2.1.tar.gz \
+    && python3 -m pip install --no-deps --ignore-installed ./freud-v2.0.0 \
+    && rm -rf freud-v2.0.0 \
+    && rm freud-v2.0.0.tar.gz \
     || exit 1
 
 
 
- curl -sSLO https://glotzerlab.engin.umich.edu/downloads/garnett/garnett-v0.4.1.tar.gz \
-    && echo "8e4860f676f92bc252d8ebc915c0ba6aeb35156f0105a685aebeb19ae54955b4  garnett-v0.4.1.tar.gz" | sha256sum -c - \
-    && tar -xzf garnett-v0.4.1.tar.gz -C . \
-    && rm -f garnett-v0.4.1/*.toml \
+ curl -sSLO https://glotzerlab.engin.umich.edu/downloads/garnett/garnett-v0.6.1.tar.gz \
+    && echo "47d4c59d2042bdb8ded415e8a5782d50777c112196070401d744df494ac3e134  garnett-v0.6.1.tar.gz" | sha256sum -c - \
+    && tar -xzf garnett-v0.6.1.tar.gz -C . \
+    && rm -f garnett-v0.6.1/*.toml \
     && export CFLAGS="-mcpu=power9 -mtune=power9" CXXFLAGS="-mcpu=power9 -mtune=power9" \
-    && python3 -m pip install --no-deps --ignore-installed ./garnett-v0.4.1 \
-    && rm -rf garnett-v0.4.1 \
-    && rm garnett-v0.4.1.tar.gz \
+    && python3 -m pip install --no-deps --ignore-installed ./garnett-v0.6.1 \
+    && rm -rf garnett-v0.6.1 \
+    && rm garnett-v0.6.1.tar.gz \
     || exit 1
 
- curl -sSLO https://glotzerlab.engin.umich.edu/Downloads/gsd/gsd-v1.8.0.tar.gz \
-    && echo "f6a99e86198a2f01e9a33b6d5eafed5ed82fa7084e1661316ac06a660e2f809d  gsd-v1.8.0.tar.gz" | sha256sum -c - \
-    && tar -xzf gsd-v1.8.0.tar.gz -C . \
-    && cd gsd-v1.8.0 \
-    && mkdir build \
+ curl -sSLO https://glotzerlab.engin.umich.edu/Downloads/gsd/gsd-v1.9.3.tar.gz \
+    && echo "89db85565b87d7a167ecf01881382ec4bc7280f07fe0295c2cf74af5db949231  gsd-v1.9.3.tar.gz" | sha256sum -c - \
+    && tar -xzf gsd-v1.9.3.tar.gz -C . \
+    && cd gsd-v1.9.3 \
+    && mkdir -p build \
     && cd build \
     && export CFLAGS="-mcpu=power9 -mtune=power9" CXXFLAGS="-mcpu=power9 -mtune=power9" \
     && cmake ../ -DPYTHON_EXECUTABLE="`which python3`" -DCMAKE_INSTALL_PREFIX=`python3 -c "import site; print(site.getsitepackages()[0])"` \
     && make install -j20 \
     && cd ../../ \
-    && rm -rf gsd-v1.8.0 \
-    && rm gsd-v1.8.0.tar.gz \
+    && rm -rf gsd-v1.9.3 \
+    && rm gsd-v1.9.3.tar.gz \
     || exit 1
 
  curl -sSLO https://glotzerlab.engin.umich.edu/Downloads/libgetar/libgetar-v1.0.1.tar.gz \
@@ -116,22 +117,22 @@ curl -sSLO https://github.com/scipy/scipy/releases/download/v1.2.0/scipy-1.2.0.t
     && rm libgetar-v1.0.1.tar.gz \
     || exit 1
 
- curl -sSLO https://glotzerlab.engin.umich.edu/Downloads/rowan/rowan-v1.2.1.tar.gz \
-    && echo "c59ad2dded288e5626db19869ed2915852734520347f7e0afd9dd08171d4541b  rowan-v1.2.1.tar.gz" | sha256sum -c - \
-    && tar -xzf rowan-v1.2.1.tar.gz -C . \
+ curl -sSLO https://glotzerlab.engin.umich.edu/Downloads/rowan/rowan-v1.2.2.tar.gz \
+    && echo "88ed2c0003543d3154374f741a04fcb21cce32180e7b29152dc93c4c887e0c51  rowan-v1.2.2.tar.gz" | sha256sum -c - \
+    && tar -xzf rowan-v1.2.2.tar.gz -C . \
     && export CFLAGS="-mcpu=power9 -mtune=power9" CXXFLAGS="-mcpu=power9 -mtune=power9" \
-    && python3 -m pip install --no-deps --ignore-installed ./rowan-v1.2.1 \
-    && rm -rf rowan-v1.2.1 \
-    && rm rowan-v1.2.1.tar.gz \
+    && python3 -m pip install --no-deps --ignore-installed ./rowan-v1.2.2 \
+    && rm -rf rowan-v1.2.2 \
+    && rm rowan-v1.2.2.tar.gz \
     || exit 1
 
- curl -sSLO https://glotzerlab.engin.umich.edu/Downloads/plato/plato-v1.6.0.tar.gz \
-    && echo "d7282e2b7a8cd641e41a845f1a7075031b821440ab1ea9ad11a0dc991b86f91d  plato-v1.6.0.tar.gz" | sha256sum -c - \
-    && tar -xzf plato-v1.6.0.tar.gz -C . \
+ curl -sSLO https://glotzerlab.engin.umich.edu/Downloads/plato/plato-v1.7.0.tar.gz \
+    && echo "8268ab6982f9135d1f7788309804eb15a1e98ef1d2f6d959733a3e23a1595def  plato-v1.7.0.tar.gz" | sha256sum -c - \
+    && tar -xzf plato-v1.7.0.tar.gz -C . \
     && export CFLAGS="-mcpu=power9 -mtune=power9" CXXFLAGS="-mcpu=power9 -mtune=power9" \
-    && python3 -m pip install --no-deps --ignore-installed ./plato-v1.6.0 \
-    && rm -rf plato-v1.6.0 \
-    && rm plato-v1.6.0.tar.gz \
+    && python3 -m pip install --no-deps --ignore-installed ./plato-v1.7.0 \
+    && rm -rf plato-v1.7.0 \
+    && rm plato-v1.7.0.tar.gz \
     || exit 1
 
  curl -sSLO https://glotzerlab.engin.umich.edu/Downloads/pythia/pythia-v0.2.5.tar.gz \
@@ -152,26 +153,26 @@ curl -sSLO https://github.com/scipy/scipy/releases/download/v1.2.0/scipy-1.2.0.t
     && rm signac-v1.2.0.tar.gz \
     || exit 1
 
- curl -sSLO https://glotzerlab.engin.umich.edu/Downloads/signac-flow/signac-flow-v0.7.1.tar.gz \
-    && echo "970ea990d8c86143161670e7c260a66357323a55ee8a5b8d8d30fe00386a929c  signac-flow-v0.7.1.tar.gz" | sha256sum -c - \
-    && tar -xzf signac-flow-v0.7.1.tar.gz -C . \
+ curl -sSLO https://glotzerlab.engin.umich.edu/Downloads/signac-flow/signac-flow-v0.8.0.tar.gz \
+    && echo "ae933964e2786f22963f70ff147a2cbd152e7e32c903ba550154b91db890451b  signac-flow-v0.8.0.tar.gz" | sha256sum -c - \
+    && tar -xzf signac-flow-v0.8.0.tar.gz -C . \
     && export CFLAGS="-mcpu=power9 -mtune=power9" CXXFLAGS="-mcpu=power9 -mtune=power9" \
-    && python3 -m pip install --no-deps --ignore-installed ./signac-flow-v0.7.1 \
-    && rm -rf signac-flow-v0.7.1 \
-    && rm signac-flow-v0.7.1.tar.gz \
+    && python3 -m pip install --no-deps --ignore-installed ./signac-flow-v0.8.0 \
+    && rm -rf signac-flow-v0.8.0 \
+    && rm signac-flow-v0.8.0.tar.gz \
     || exit 1
 
- curl -sSLO https://glotzerlab.engin.umich.edu/Downloads/hoomd/hoomd-v2.6.0.tar.gz \
-    && echo "f926b2d39cf58916794edbf4c619d0542a9b1e66b6c6a1b6a01b70f5617ec728  hoomd-v2.6.0.tar.gz" | sha256sum -c - \
-    && tar -xzf hoomd-v2.6.0.tar.gz -C . \
-    && cd hoomd-v2.6.0 \
-    && mkdir build \
+ curl -sSLO https://glotzerlab.engin.umich.edu/Downloads/hoomd/hoomd-v2.8.0.tar.gz \
+    && echo "f3cd733bb9a8932a30d65927897eaff894bf5ddb7a95c507ee47e124a7a4f541  hoomd-v2.8.0.tar.gz" | sha256sum -c - \
+    && tar -xzf hoomd-v2.8.0.tar.gz -C . \
+    && cd hoomd-v2.8.0 \
+    && mkdir -p build \
     && cd build \
     && export CFLAGS="-mcpu=power9 -mtune=power9" CXXFLAGS="-mcpu=power9 -mtune=power9" \
     && cmake ../ -DENABLE_CUDA=on -DENABLE_MPI=on -DENABLE_TBB=on -DBUILD_JIT=on -DBUILD_TESTING=off -DCMAKE_INSTALL_PREFIX=`python3 -c "import site; print(site.getsitepackages()[0])"` \
     && make install -j20 \
     && cd ../../ \
-    && rm -rf /root/hoomd-v2.6.0 \
-    && rm hoomd-v2.6.0.tar.gz \
+    && rm -rf /root/hoomd-v2.8.0 \
+    && rm hoomd-v2.8.0.tar.gz \
     || exit 1